CU-1wm6eay_Search-Component_Sanja-Nikolic, moved search bar to ui folder
