+++ frontend/src/components/UI/searchBar/SearchBar.test.tsx
