+++ frontend/src/components/UI/searchBar/searchBar-styles.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { theme } from '../../theme';
+import { theme } from '../../../theme';
 
 export const SearchInputField = styled.input`
   height: ${theme.space.lg};
